Post Fix Expression Decoder
This project reads in a post fix expression from a binary file and
pushes the items to a stack.  When a non number is read it pops the top
two numbers, evaluates the expression, and pushes the answer back to
the stack.  This is done until a final solution is found.